+++ src/main/java/net/kaczmarzyk/spring/data/jpa/utils/Converter.java
@@ -136,6 +136,8 @@ public class Converter {
 			return (T) convertToEnum(value, (Class<? extends Enum<?>>) expectedClass, ignoreCase);
 		} else if (expectedClass.isAssignableFrom(Date.class)) {
 			return (T) convertToDate(value);
+		} else if (expectedClass.isAssignableFrom(Calendar.class)) {
+			return (T) convertToCalendar(value);
 		} else if (isAssignableFromAnyOf(expectedClass, Boolean.class, boolean.class)) {
 			return (T) convertToBoolean(value);
 		} else if (isAssignableFromAnyOf(expectedClass, Integer.class, int.class, Long.class, long.class)) {
@@ -238,6 +240,17 @@ public class Converter {
 		String dateFormat = getDateFormat(Date.class);
 		try {
 			return new SimpleDateFormat(dateFormat).parse(value);
+		} catch (ParseException e) {
+			throw new ValueRejectedException(value, "Date format exception, expected format: " + dateFormat);
+		}
+	}
+
+	public Calendar convertToCalendar(String value) {
+		String dateFormat = getDateFormat(Date.class);
+		try {
+			Calendar cal = Calendar.getInstance();
+			cal.setTime(new SimpleDateFormat(dateFormat).parse(value));
+			return cal;
 		} catch (ParseException e) {
 			throw new ValueRejectedException(value, "Date format exception, expected format: " + dateFormat);
 		}