+++ src/test/java/net/kaczmarzyk/JoinE2eTest.java
@@ -22,6 +22,8 @@ import net.kaczmarzyk.spring.data.jpa.domain.In;
 import net.kaczmarzyk.spring.data.jpa.domain.Like;
 import net.kaczmarzyk.spring.data.jpa.domain.LikeIgnoreCase;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.*;
+import net.kaczmarzyk.utils.interceptor.HibernateStatementInterceptor;
+import static net.kaczmarzyk.utils.interceptor.InterceptedStatementsAssert.assertThatInterceptedStatements;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -77,7 +79,7 @@ public class JoinE2eTest extends E2eTestBase {
 
 		@Join(path = "orders", alias = "o")
 		@Spec(path = "o.itemName", params = "orderIn", spec = In.class)
-		public interface OrderInSpecification extends Specification<Customer> {
+		private interface OrderInSpecification extends Specification<Customer> {
 		}
 
 		@RequestMapping(value = "/join/customers", params = { "orderIn" })
@@ -88,7 +90,7 @@ public class JoinE2eTest extends E2eTestBase {
 
 		@RequestMapping(value = "/multi-join/customers", params = { "order", "badge" })
 		@ResponseBody
-		public Object findByOrderAndOrders2(
+		public Object findByOrdersAndBadges(
 
 				@Joins({
 						@Join(path = "orders", alias = "o"),
@@ -193,6 +195,25 @@ public class JoinE2eTest extends E2eTestBase {
 			.andExpect(jsonPath("$.totalPages").value(2))
 			.andExpect(jsonPath("$.totalElements").value(2))
 			.andExpect(jsonPath("$.size").value(1));
+	}
+
+	@Test
+	public void reusesEvaluatedJoinForManySpecs() throws Exception {
+		em.flush();
+		HibernateStatementInterceptor.clearInterceptedStatements();
+
+		mockMvc.perform(get("/join/customers")
+				.param("order1", "Beer")
+				.param("order2", "Donuts")
+				.accept(MediaType.APPLICATION_JSON))
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("$").isArray())
+			.andExpect(jsonPath("$[0].firstName").value("Homer"))
+			.andExpect(jsonPath("$[1].firstName").value("Moe"))
+			.andExpect(jsonPath("$[2]").doesNotExist());
+
+		assertThatInterceptedStatements()
+				.hasSingleSelectWithNumberOfJoins(1);
 	}
 
 }