+++ src/test/java/net/kaczmarzyk/spring/data/jpa/CustomerRepositoryWithCache.java
@@ -13,29 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.kaczmarzyk.spring.data.jpa.utils;
+package net.kaczmarzyk.spring.data.jpa;
 
-import javax.persistence.criteria.Fetch;
-import javax.persistence.criteria.Join;
-import javax.persistence.criteria.Root;
-import java.util.function.Function;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 /**
- * Ugly way to share context between different specifications -- e.g. joins (see
- * {@code JoinSpecificationResolver})
- *
  * @author Tomasz Kaczmarzyk
  */
-public interface QueryContext {
+@Repository
+public interface CustomerRepositoryWithCache extends PagingAndSortingRepository<Customer, Long>, JpaSpecificationExecutor<Customer> {
 
-    boolean existsJoin(String key, Root<?> root);
-
-    Join<?, ?> getEvaluated(String key, Root<?> root);
-
-    void putLazyVal(String key, Function<Root<?>, Join<?, ?>> value);
-
-    Fetch<?, ?> getEvaluatedJoinFetch(String key);
-
-    void putEvaluatedJoinFetch(String key, Fetch<?, ?> fetch);
-
+	@Cacheable("customers")
+	List<Customer> findAll(Specification<Customer> spec);
 }