+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/BetweenTest.java
@@ -20,6 +20,8 @@ import net.kaczmarzyk.spring.data.jpa.Gender;
 import net.kaczmarzyk.spring.data.jpa.IntegrationTestBase;
 import net.kaczmarzyk.spring.data.jpa.utils.Converter;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.OnTypeMismatch;
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -188,4 +190,12 @@ public class BetweenTest extends IntegrationTestBase {
     public void rejectsTooManyArguments() throws ParseException {
         new Between<>(queryCtx, "path", new String[] { "2014-03-10", "2014-03-11", "2014-03-11" }, defaultConverter);
     }
+
+	@Test
+	public void equalsAndHashCodeContract() {
+		EqualsVerifier.forClass(Between.class)
+				.usingGetClass()
+				.suppress(Warning.NONFINAL_FIELDS)
+				.verify();
+	}
 }